# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Initiative Tracker API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /initiatives:
    get:
      tags: [Initiatives]
      operationId: listInitiatives
      summary: List all initiatives
      responses:
        '200':
          description: List of initiatives with progress.
          content:
            application/json:
              schema:
                type: object
                required: [initiatives]
                properties:
                  initiatives:
                    type: array
                    items:
                      $ref: '#/components/schemas/InitiativeWithProgress'
    post:
      tags: [Initiatives]
      operationId: createInitiative
      summary: Create a new initiative (project-manager or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInitiativeRequest'
      responses:
        '201':
          description: Initiative created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Initiative'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /initiatives/{id}:
    get:
      tags: [Initiatives]
      operationId: getInitiative
      summary: Get a single initiative with milestones
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Initiative detail with milestones.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiativeDetail'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      tags: [Initiatives]
      operationId: updateInitiative
      summary: Update an initiative (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInitiativeRequest'
      responses:
        '200':
          description: Initiative updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Initiative'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Initiatives]
      operationId: deleteInitiative
      summary: Delete an initiative (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Initiative:
      type: object
      required: [id, title, description, ownerId, status, startDate, targetDate, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: [string, 'null']
        ownerId:
          type: string
        ownerName:
          type: [string, 'null']
        status:
          type: string
          enum: [draft, active, completed, on_hold, cancelled]
        priority:
          type: [string, 'null']
          enum: [low, medium, high, critical, null]
        startDate:
          type: [string, 'null']
          format: date
        targetDate:
          type: [string, 'null']
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InitiativeWithProgress:
      allOf:
        - $ref: '#/components/schemas/Initiative'
        - type: object
          required: [progress, milestoneCount, approvedCount, pendingCount]
          properties:
            progress:
              type: number
              minimum: 0
              maximum: 100
            milestoneCount:
              type: integer
            approvedCount:
              type: integer
            pendingCount:
              type: integer
    InitiativeDetail:
      allOf:
        - $ref: '#/components/schemas/InitiativeWithProgress'
        - type: object
          required: [milestones]
          properties:
            milestones:
              type: array
              items:
                $ref: '#/components/schemas/MilestoneWithAttachments'
    CreateInitiativeRequest:
      type: object
      required: [title, status]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [draft, active, completed, on_hold, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        startDate:
          type: string
          format: date
        targetDate:
          type: string
          format: date
    UpdateInitiativeRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [draft, active, completed, on_hold, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        startDate:
          type: string
          format: date
        targetDate:
          type: string
          format: date
    Milestone:
      type: object
      required: [id, initiativeId, title, status, weight, createdAt, updatedAt]
      properties:
        id:
          type: integer
        initiativeId:
          type: integer
        title:
          type: string
        description:
          type: [string, 'null']
        status:
          type: string
          enum: [pending, in_progress, submitted, approved, rejected]
        weight:
          type: number
          minimum: 0
          maximum: 100
        dueDate:
          type: [string, 'null']
          format: date
        approvedById:
          type: [string, 'null']
        approvedByName:
          type: [string, 'null']
        approvedAt:
          type: [string, 'null']
          format: date-time
        rejectionReason:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MilestoneWithAttachments:
      allOf:
        - $ref: '#/components/schemas/Milestone'
        - type: object
          required: [attachments]
          properties:
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/FileAttachment'
...
    FileAttachment:
      type: object
      required: [id, milestoneId, fileName, objectPath, uploadedById, createdAt]
      properties:
        id:
          type: integer
        milestoneId:
          type: integer
        fileName:
          type: string
        contentType:
          type: [string, 'null']
        objectPath:
          type: string
        uploadedById:
          type: string
        uploadedByName:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...