# lib/api-zod/src/generated/types/milestoneWithAttachments.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileAttachment } from "./fileAttachment";
import type { Milestone } from "./milestone";

export type MilestoneWithAttachments = Milestone & {
  attachments: FileAttachment[];
};