# lib/api-zod/src/generated/types/updateInitiativeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateInitiativeRequestPriority } from "./updateInitiativeRequestPriority";
import type { UpdateInitiativeRequestStatus } from "./updateInitiativeRequestStatus";

export interface UpdateInitiativeRequest {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: UpdateInitiativeRequestStatus;
  priority?: UpdateInitiativeRequestPriority;
  startDate?: Date;
  targetDate?: Date;
}